GH Actions: don't fail fast on "low" PHP versions
I'd like to suggest adding this as a temporary measure until the failing code has been fixed.

The `fail-fast` setting defaults to `true` and has the effect of cancelling any and all running/waiting builds within the same workflow as soon as one build has failed.

By setting this to `false`, all builds will be always be run, allowing for easier debugging of the current test failures.