Fix variable names to match ext/sodium
This is technically an API break for PHP 8 projects.
However, this polyfill always targeted PHP 5, 7.0, and 7.1. You shouldn't need it on 7.2+.

The end result of this change is that you should be able to write code that uses named parameters on PHP 8 and have it work regardless of ext/sodium being installed or not (provided you import sodium_compat via composer).

Additionally, we include lib/ in the psalm checks (if applicable).

This will need a new minor version (v1.19.0).