Update Wycheproof tests from C2SP
Project Wycheproof is now part of C2SP. This pulls in the latest test vectors for ChaCha20-Poly1305, XChaCha20-Poly1305, and X25519. Additionally, we add the SipHash-2-4 tests from Wycheproof.

The X25519 tests still take several minutes to run on a decent machine, so they are still gated behind the pedantic test flag.