Optimize curve25519 field element by removing array
By leveraging PHP's optimized object key access for what is effectively always a fixed-length array, we can remove some computational overhead and memory usage in by the PHP interpreter.

Since these memory access patterns are not secret (they were array indices from 0 to 9), the risk of side-channel leakage by this change is at least as low as what was previously implemented.

This is a backport of #198 to the v1.x branch. It's not a 1:1 copy, as some code style changes occur in PHP 8.1+.